:truck: chore: Rename files in shared/components as requested in the issue
