+++ src/app/modules/shared/components/select/select.css
