+++ src/app/modules/shared/components/select/select.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SelectComponent } from './select.component';
+import { SelectComponent } from './select';
 
 describe('SelectComponent', () => {
   let component: SelectComponent;
