+++ src/app/modules/shared/shared-module.ts
@@ -12,8 +12,7 @@ import { Login } from './pages/login/login';
 import { Table } from './components/table/table';
 import { CheckboxComponent } from './components/checkbox/checkbox';
 import { PageLayout } from './components/page-layout/page-layout';
-import { SelectComponent } from './components/select/select.component';
-
+import { SelectComponent } from './components/select/select';
 
 @NgModule({
   declarations: [
